requirements-api.txt
aioredis>=2.0.1
aiofiles>=25.1.0
aiohttp==3.10.5
aiosqlite>=0.21.0
alembic>=1.17.2
asyncpg>=0.29.0
awscli>=1.43.7
beautifulsoup4==4.12.3
boto3>=1.42.0
chardet>=5.2.0
curl-cffi>=0.13.0
fake-useragent>=2.2.0
fastapi==0.115.0
greenlet>=3.2.4
httpx<0.28
langchain>=1.1.0
langchain-core>=1.0.3
langchain-text-splitters>=0.3.0
langgraph>=1.0.2
lingua-language-detector==2.1.1
loguru==0.7.2
markdown>=3.10
networkx>=3.0
numpy==2.1.2
openai>=1.45.0
pandas==2.2.3
pdfplumber>=0.9.0
pillow>=10.0.0
psycopg2-binary>=2.9.11
pydantic-settings>=2.11.0
pymupdf>=1.26.6
pypdf>=4.0.0
pypdfium2>=4.30.0
pytesseract>=0.3.13
python-dateutil>=2.9.0.post0
python-dotenv>=1.2.1
python-multipart>=0.0.20
pyjwt>=2.10.1
qdrant-client>=1.15.1
rank-bm25>=0.2.2
redis>=5.0.0
requests==2.32.3
selenium>=4.38.0
sentence-transformers==3.0.1
sqlalchemy==2.0.36
tavily-python>=0.7.13
tiktoken>=0.8.0
torch>=2.6,<3.0
tornado<6.4
uvicorn==0.31.1
webdriver-manager>=4.0.2
weasyprint>=67.0
